[Truffle] Fix order of specializations for RangeLiteralNode.

+++ core/src/main/java/org/jruby/truffle/nodes/literal/RangeLiteralNode.java
@@ -49,12 +49,12 @@ public abstract class RangeLiteralNode extends RubyNode {
     }
 
     @Specialization
-    public RubyRange.LongFixnumRange doRange(long begin, long end) {
+    public RubyRange.LongFixnumRange doRange(long begin, int end) {
         return new RubyRange.LongFixnumRange(getContext().getCoreLibrary().getRangeClass(), begin, end, excludeEnd);
     }
 
     @Specialization
-    public RubyRange.LongFixnumRange doRange(long begin, int end) {
+    public RubyRange.LongFixnumRange doRange(long begin, long end) {
         return new RubyRange.LongFixnumRange(getContext().getCoreLibrary().getRangeClass(), begin, end, excludeEnd);
     }
